requirements.txt
acres==0.2.0
altair==5.5.0
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
cryptography==44.0.0
etelemetry==0.3.1
filelock==3.16.1
gitdb==4.0.12
GitPython==3.1.44
httplib2==0.22.0
idna==3.10
importlib_resources==6.5.2
isodate==0.6.1
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Levenshtein==0.26.1
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.22.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20231228
pillow==11.1.0
protobuf==5.29.3
prov==2.0.1
puremagic==1.28
pyarrow==19.0.0
pycparser==2.22
pydeck==0.9.1
pydot==3.0.4
Pygments==2.19.1
PyMuPDF==1.25.1
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
pytz==2024.2
pyxnat==1.6.2
RapidFuzz==3.11.0
rdflib==6.3.2
referencing==0.36.1
reportlab==4.2.5
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scipy==1.15.1
simplejson==3.19.3
six==1.17.0
smmap==5.0.2
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
traits==6.4.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
watchdog==6.0.0
